Rename contacts to idprovider- packages & files
Rename contacts to idprovider, contactsyaml to local, contactstest to
idprovidertest and files yamlprovider.go to idprovider.go and
provider.go to idprovider.go
