+++ idprovider/idprovidertest/idprovider.go
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package contactstest
+package idprovidertest
 
 import (
 	"io/ioutil"
