+++ idprovider/local/idprovider.go
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package contactsyaml
+package local
 
 import (
 	"fmt"
